August 17,2019
-Trying to do some tweaks to one of my old project. Still
need to design a logo and adjust some of the colors plus some 
some parts such as the footer and the pop
up window I would still need to adjust some of the layout.
Wondering what does this security vulnerabilities mean in 
my dependencies going to need to read more about them. 

Planning to recreate this from scratch so that I could
remember and retain how I made it using SASS and use of the BEM 
methodology of naming classes. 

Change row class col color to white